Another batch of removed deprecated code
