+++ app/bundles/CampaignBundle/Entity/EventRepository.php
@@ -11,10 +11,9 @@
 
 namespace Mautic\CampaignBundle\Entity;
 
-/**
- * EventRepository.
- */
-class EventRepository extends LegacyEventRepository
+use Mautic\CoreBundle\Entity\CommonRepository;
+
+class EventRepository extends CommonRepository
 {
     /**
      * Get a list of entities.
@@ -289,6 +288,34 @@ class EventRepository extends LegacyEventRepository
         $results = $q->getQuery()->getResult();
 
         return $results;
+    }
+
+    /**
+     * Get an array of events that have been triggered by this lead.
+     *
+     * @param $leadId
+     *
+     * @return array
+     */
+    public function getLeadTriggeredEvents($leadId)
+    {
+        $q = $this->getEntityManager()->createQueryBuilder()
+            ->select('e, c, l')
+            ->from('MauticCampaignBundle:Event', 'e')
+            ->join('e.campaign', 'c')
+            ->join('e.log', 'l');
+
+        //make sure the published up and down dates are good
+        $q->where($q->expr()->eq('IDENTITY(l.lead)', (int) $leadId));
+
+        $results = $q->getQuery()->getArrayResult();
+
+        $return = [];
+        foreach ($results as $r) {
+            $return[$r['id']] = $r;
+        }
+
+        return $return;
     }
 
     /**
