+++ app/bundles/CampaignBundle/Event/CampaignBuilderEvent.php
@@ -16,9 +16,6 @@ use Mautic\CoreBundle\Event\ComponentValidationTrait;
 use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\Translation\TranslatorInterface;
 
-/**
- * Class CampaignBuilderEvent.
- */
 class CampaignBuilderEvent extends Event
 {
     use ComponentValidationTrait;
@@ -98,33 +95,11 @@ class CampaignBuilderEvent extends Event
     }
 
     /**
-     * Get decisions.
-     *
      * @return mixed
      */
     public function getDecisions()
     {
         return $this->sort('decisions');
-    }
-
-    /**
-     * @deprecated - use addDecision instead
-     *
-     * @param $key
-     */
-    public function addLeadDecision($key, array $decision)
-    {
-        $this->addDecision($key, $decision);
-    }
-
-    /**
-     * @deprecated - use getDecisions instead
-     *
-     * @return array
-     */
-    public function getLeadDecisions()
-    {
-        return $this->getDecisions();
     }
 
     /**
@@ -174,26 +149,6 @@ class CampaignBuilderEvent extends Event
     public function getConditions()
     {
         return $this->sort('conditions');
-    }
-
-    /**
-     * @deprecated use addCondition instead
-     *
-     * @param $key
-     */
-    public function addLeadCondition($key, array $event)
-    {
-        $this->addCondition($key, $event);
-    }
-
-    /**
-     * @deprecated use getConditions() instead
-     *
-     * @return array
-     */
-    public function getLeadConditions()
-    {
-        return $this->getConditions();
     }
 
     /**
